+++ Desktop/IB/security/src/main/java/uns/ac/rs/ib/security/repository/HealthSerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import uns.ac.rs.ib.security.model.HealthSer;
 
 @Repository
-public interface HealthSerRepository extends JpaRepository<HealthSer, Long> {
+public interface HealthSerRepository extends JpaRepository<HealthSer, Integer> {
 
 }
