+++ Desktop/IB/security/src/main/java/uns/ac/rs/ib/security/repository/MedicalRecordRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import uns.ac.rs.ib.security.model.MedicalRecord;
 
 @Repository
-public interface MedicalRecordRepository extends JpaRepository<MedicalRecord, Long> {
+public interface MedicalRecordRepository extends JpaRepository<MedicalRecord, Integer> {
 
 }
